Align campaign audience position type

+++ supabase/migrations/20260904134500_prospecting_campaign_audience_search.sql
@@ -167,7 +167,7 @@ BEGIN
     member.timezone,
     member.status,
     member.suppression_reason,
-    member.current_step_position,
+    member.current_step_position::integer,
     member.next_action_at,
     member.enrolled_at,
     lead.full_name,
